requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.21.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.62.0
PyYAML>=6.0
click>=8.0.0
pillow>=9.0.0
networkx>=2.8.0

# Optional dependencies for development
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
mypy>=1.0.0
pre-commit>=2.17.0

# Optional dependencies for experiment tracking
mlflow>=2.0.0
wandb>=0.13.0